requirements.txt
groq>=0.4.0
pyyaml>=6.0
psycopg2-binary>=2.9.0
